requirements.txt
requests
python-dotenv
huggingface_hub